Add files via upload
Amazon Clone Project

This project is a simplified clone of the Amazon e-commerce website, built using HTML and CSS. It aims to replicate the basic user interface and layout of Amazon's homepage, including key features such as a responsive navigation bar, product listings, and a footer. This project serves as a practical exercise in web development, focusing on structuring web pages with HTML and styling them with CSS to create a visually appealing and user-friendly experience.